Orphan-resources/azure-orphan-detector/requirements.txt
# Development dependencies
azure-identity>=1.12.0
azure-mgmt-resource>=21.0.0
azure-mgmt-compute>=29.0.0
azure-mgmt-network>=22.0.0
azure-mgmt-storage>=20.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-costmanagement>=4.0.0
azure-mgmt-recoveryservicesbackup>=4.1.0
azure-mgmt-subscription>=3.1.1
jinja2>=3.1.0
pyyaml>=6.0
rich>=12.0.0
typer>=0.7.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
isort>=5.10.0
pre-commit>=2.20.0

# Documentation
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0